requirements.txt
flask
requests
python-dotenv
supabase
sendgrid